Simplify Fibonacci code using negative indices
Used [-1] and [-2] to reference the last and second to last elements in
the array, instead of using the array length -1/-2. This applied to both
methods. For fibs() this meant the fib_num was not needed, meaning the
initial enumerable could be simplified to (num - 2).times.

I also realized that the num == 2 guard clause in fibs() is not required
since it is covered by the reduce returning the initial value of [0, 1]
when the loop does not start at all (when num == 2).